scheduler now fully functional, webpages reflect update rate, webpage reloads every 2 secs.

+++ src/github.com/gonium/gosdm630/scheduler.go
@@ -26,25 +26,25 @@ func NewRoundRobinScheduler(
 func (s *RoundRobinScheduler) Produce() {
 	for {
 		for _, devid := range s.devids {
-			s.out <- QuerySnip{DeviceId: devid, OpCode: OpCodeL1Voltage, Value: float32(math.NaN())}
-			s.out <- QuerySnip{DeviceId: devid, OpCode: OpCodeL2Voltage, Value: float32(math.NaN())}
-			s.out <- QuerySnip{DeviceId: devid, OpCode: OpCodeL3Voltage, Value: float32(math.NaN())}
+			s.out <- QuerySnip{DeviceId: devid, OpCode: OpCodeL1Voltage, Value: math.NaN()}
+			s.out <- QuerySnip{DeviceId: devid, OpCode: OpCodeL2Voltage, Value: math.NaN()}
+			s.out <- QuerySnip{DeviceId: devid, OpCode: OpCodeL3Voltage, Value: math.NaN()}
 
-			s.out <- QuerySnip{DeviceId: devid, OpCode: OpCodeL1Current, Value: float32(math.NaN())}
-			s.out <- QuerySnip{DeviceId: devid, OpCode: OpCodeL2Current, Value: float32(math.NaN())}
-			s.out <- QuerySnip{DeviceId: devid, OpCode: OpCodeL3Current, Value: float32(math.NaN())}
+			s.out <- QuerySnip{DeviceId: devid, OpCode: OpCodeL1Current, Value: math.NaN()}
+			s.out <- QuerySnip{DeviceId: devid, OpCode: OpCodeL2Current, Value: math.NaN()}
+			s.out <- QuerySnip{DeviceId: devid, OpCode: OpCodeL3Current, Value: math.NaN()}
 
-			s.out <- QuerySnip{DeviceId: devid, OpCode: OpCodeL1Cosphi, Value: float32(math.NaN())}
-			s.out <- QuerySnip{DeviceId: devid, OpCode: OpCodeL2Cosphi, Value: float32(math.NaN())}
-			s.out <- QuerySnip{DeviceId: devid, OpCode: OpCodeL3Cosphi, Value: float32(math.NaN())}
+			s.out <- QuerySnip{DeviceId: devid, OpCode: OpCodeL1Cosphi, Value: math.NaN()}
+			s.out <- QuerySnip{DeviceId: devid, OpCode: OpCodeL2Cosphi, Value: math.NaN()}
+			s.out <- QuerySnip{DeviceId: devid, OpCode: OpCodeL3Cosphi, Value: math.NaN()}
 
-			s.out <- QuerySnip{DeviceId: devid, OpCode: OpCodeL1Import, Value: float32(math.NaN())}
-			s.out <- QuerySnip{DeviceId: devid, OpCode: OpCodeL2Import, Value: float32(math.NaN())}
-			s.out <- QuerySnip{DeviceId: devid, OpCode: OpCodeL3Import, Value: float32(math.NaN())}
+			s.out <- QuerySnip{DeviceId: devid, OpCode: OpCodeL1Import, Value: math.NaN()}
+			s.out <- QuerySnip{DeviceId: devid, OpCode: OpCodeL2Import, Value: math.NaN()}
+			s.out <- QuerySnip{DeviceId: devid, OpCode: OpCodeL3Import, Value: math.NaN()}
 
-			s.out <- QuerySnip{DeviceId: devid, OpCode: OpCodeL1Export, Value: float32(math.NaN())}
-			s.out <- QuerySnip{DeviceId: devid, OpCode: OpCodeL2Export, Value: float32(math.NaN())}
-			s.out <- QuerySnip{DeviceId: devid, OpCode: OpCodeL3Export, Value: float32(math.NaN())}
+			s.out <- QuerySnip{DeviceId: devid, OpCode: OpCodeL1Export, Value: math.NaN()}
+			s.out <- QuerySnip{DeviceId: devid, OpCode: OpCodeL2Export, Value: math.NaN()}
+			s.out <- QuerySnip{DeviceId: devid, OpCode: OpCodeL3Export, Value: math.NaN()}
 		}
 	}
 }
